requirements.txt
click==8.1.3
Flask==2.2.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.31
mysqlclient==2.1.1
protobuf==3.20.1
SQLAlchemy==1.4.42
Werkzeug==2.2.2
